別でも使いたいのでsubscribeTimerとして分離して、useTimerを解体
